[Utils] Move MultiStepInput out of external dir
This commit moves MultiStepInput class out of external dir.
This commit contains the code chage about MultiStepInput.

ONE-vscode-DCO-1.0-Signed-off-by: Jiyoung Yun <jy910.yun@samsung.com>
