+++ src/Utils/MultiStepInput.ts
@@ -1,3 +1,18 @@
+/*
+ * Copyright (c) 2022 Samsung Electronics Co., Ltd. All Rights Reserved
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 /*---------------------------------------------------------------------------------------------
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
@@ -46,7 +61,12 @@ export class MultiStepInput {
   }
 
   private current?: QuickInput;
-  private steps: InputStep[] = [];
+
+  // NOTE(jyoung)
+  // When a new quick input is shown in the progress of using MultiStepInput, it is
+  // automatically added to this steps variable. To remove unintended quick input,
+  // this variable sets to public.
+  public steps: InputStep[] = [];
 
   private async stepThrough<T>(start: InputStep) {
     let step: InputStep|void = start;
